fixed about 200 warnings on Win64
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.4@5016 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FILTERING/SMOOTHING/SavitzkyGolayFilter.h
@@ -112,7 +112,7 @@ namespace OpenMS
       template <typename PeakType>
       void filter(MSSpectrum<PeakType>& spectrum)
       {
-        UInt n = spectrum.size();
+        UInt n = (UInt)spectrum.size();
         
         typename MSSpectrum<PeakType>::iterator first = spectrum.begin();
         typename MSSpectrum<PeakType>::iterator last = spectrum.end();
